# shared/openapi.yaml
openapi: 3.0.3
info:
  title: National Zoning & Land Use Data Platform API
  version: 0.1.0
  description: |
    REST API for municipalities, documents, structured zoning data, RAG, and jobs.
    Expand this spec as endpoints are implemented; optionally generate a TypeScript client for the frontend.
servers:
  - url: http://localhost:5000
    description: Local Flask
paths:
...
  /api/v1/zones/{zone_id}:
    get:
      summary: Single zone record (GeoJSON geometry when available)
      operationId: getZoneById
      parameters:
        - name: zone_id
          in: path
          required: true
          schema:
            type: integer
  /api/v1/zones/{zone_id}/analyze:
    post:
      summary: Ingest linked PDFs for the zone, then run RAG (default or custom question in body)
      operationId: postZoneAnalyze
      parameters:
        - name: zone_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                q:
                  type: string
                  description: Optional RAG question; omit for server default prompt
...